AutoAnnSmoother/requirements.txt
nuscenes-devkit==1.1.8
numpy==1.19.2
#open3d==0.10.0.0
matplotlib==3.1.3
opencv-python==4.3.0.38
#ninja==1.9.0.post1
ipympl
wandb
plotly
pyquaternion
#k3d==2.12.0
pandas==1.3.0
#einops==0.4.0
motmetrics==1.1.3
pandas==1.3.0
seaborn
#trimesh
numba==0.55.1
zod